fix: fix crash on market page
